refactor: lib폴더 제거하고 src로 통일
