+++ packages/ui/index.ts
@@ -1,3 +1,2 @@
-export * from './lib/components';
 export * from './src/components';
-export * from './src/lib';
+export * from './src/utils';
