--- packages/ui/lib/components/index.ts
@@ -1,5 +1,0 @@
-export { default as Toast } from './Toast';
-export { default as Loading } from './Loading';
-export { default as ErrorFallback } from './ErrorFallback';
-export { default as ErrorBoundary } from './ErrorBoundary';
-export { default as LoadingFallback } from './LoadingFallback';
