+++ packages/ui/src/components/index.ts
@@ -1,1 +1,6 @@
 export * from './button';
+export { default as Toast } from './Toast';
+export { default as Loading } from './Loading';
+export { default as ErrorFallback } from './ErrorFallback';
+export { default as ErrorBoundary } from './ErrorBoundary';
+export { default as LoadingFallback } from './LoadingFallback';
