+++ packages/ui/src/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './utils';
+export * from './cn';
 export * from './withUI';
